tests: update testing structure
